[feat] Add support for ViLT cmd pretraining
Add yaml configs and joint head for cmd pretraining with megaVLT.

[ghstack-poisoned]
